+++ items/help/help_missing_keys.sh
@@ -20,7 +20,7 @@ static_content() {
         0.0 T "-#[nodim]not be available with the current" \
         0.0 T "-#[nodim]keyboard settings."
 
-    if $cfg_use_whiptail; then
+    if $b_use_alt_handler; then
         set -- "$@" \
             0.0 T " " \
             0.0 T "When using whiptail it is not possible" \