+++ items/odd_chars/currencies.sh
@@ -11,7 +11,7 @@
 
 show_label() {
     # Some Currency symbols can't be printed in whiptail
-    if $cfg_use_whiptail; then
+    if $b_use_alt_handler; then
         case "$1" in
             ₺ | ₴ | ₽ | ₹ | ₿)
                 echo "Send   ($2) - not printable in whiptail"
@@ -75,7 +75,7 @@ no_auto_menu_handling=1 # delay processing of dialog, only source it for now
 
 if [ -n "$1" ]; then
     "$D_TM_BASE_PATH"/scripts/act_display_char.sh "$1"
-elif $cfg_use_whiptail; then
+elif $b_use_alt_handler; then
     tmux_error_handler set-option -gqu "$wt_pasting"
 fi