+++ items/odd_chars/diacritics_e.sh
@@ -55,7 +55,7 @@ D_TM_BASE_PATH=$(cd "${0%/*}/../.." && pwd)
 
 if [ -n "$1" ]; then
     "$D_TM_BASE_PATH"/scripts/act_display_char.sh "$1"
-elif $cfg_use_whiptail; then
+elif $b_use_alt_handler; then
     tmux_error_handler set-option -gqu "$wt_pasting"
 fi