+++ items/paste_buffers.sh
@@ -13,7 +13,7 @@ static_content() {
     select_cmd="$TMUX_BIN choose-buffer"
     tmux_vers_check 2.6 && select_cmd="$select_cmd -Z"
 
-    if $cfg_use_hint_overlays && ! $cfg_use_whiptail; then
+    if $cfg_use_hint_overlays && ! $b_use_alt_handler; then
         select_cmd="$select_cmd \& $d_hints/choose-buffer.sh skip-oversized"
     fi
 
@@ -25,7 +25,7 @@ static_content() {
     set -- \
         0.0 S
 
-    if ! $cfg_use_whiptail; then
+    if ! $b_use_alt_handler; then
         set -- "$@" \
             0.0 C v "Paste the most recent paste buffer" "paste-buffer -p  $runshell_reload_mnu"
     fi