+++ scripts/menu_handling.sh
@@ -70,7 +70,7 @@ starting_with_dash() {
 }
 
 is_function_defined() {
-    command -v "$1" >/dev/null 2>&1
+    command -v "$1" >/dev/null
 }
 
 run_if_found() {
@@ -117,8 +117,8 @@ verify_menu_runable() {
     # extract first word
     _actual_first="${menu_items%% *}"
 
-    if [ -n "$cfg_alt_menu_handler" ]; then
-        _mnu_first="$cfg_alt_menu_handler"
+    if [ -n "$alt_menu_handler" ]; then
+        _mnu_first="$alt_menu_handler"
     else
         _mnu_first="${TMUX_BIN%% *}"
     fi
@@ -300,11 +300,11 @@ mnu_spacer() {
 }
 
 alt_prefix() {
-    case "$cfg_alt_menu_handler" in
+    case "$alt_menu_handler" in
         whiptail | dialog) ;;
-        *) error_msg "Un-recognized cfg_alt_menu_handler: [$cfg_alt_menu_handler]" ;;
+        *) error_msg "Un-recognized alt_menu_handler: [$alt_menu_handler]" ;;
     esac
-    menu_items="$cfg_alt_menu_handler --title \"$menu_name\"  --menu \"\" 0 0 0 "
+    menu_items="$alt_menu_handler --title \"$menu_name\"  --menu \"\" 0 0 0 "
 }
 
 alt_open_menu() {
@@ -406,7 +406,7 @@ menu_parse() {
     menu_items=""
     [ "$menu_idx" -eq 1 ] && {
         # set prefix for item 1
-        if $cfg_use_whiptail; then
+        if $b_use_alt_handler; then
             alt_prefix
         else
             mnu_prefix
@@ -439,7 +439,7 @@ menu_parse() {
 
                 [ -n "$menu_debug" ] && debug_print "key[$_mp_key] label[$_mp_label] command[$_mp_cmd]"
 
-                if $cfg_use_whiptail; then
+                if $b_use_alt_handler; then
                     alt_command "$_mp_label" "$_mp_key" "$_mp_cmd"
                 else
                     mnu_command "$_mp_label" "$_mp_key" "$_mp_cmd"
@@ -474,7 +474,7 @@ menu_parse() {
 
                 [ -n "$menu_debug" ] && debug_print "key[$_mp_key] label[$_mp_label] command[$_mp_cmd]"
 
-                if $cfg_use_whiptail; then
+                if $b_use_alt_handler; then
                     alt_external_cmd "$_mp_label" "$_mp_key" "$_mp_cmd"
                 else
                     mnu_external_cmd "$_mp_label" "$_mp_key" "$_mp_cmd"
@@ -507,7 +507,7 @@ menu_parse() {
 
                 [ -n "$menu_debug" ] && debug_print "key[$_mp_key] label[$_mp_label] menu[$menu]"
 
-                if $cfg_use_whiptail; then
+                if $b_use_alt_handler; then
                     alt_open_menu "$_mp_label" "$_mp_key" "$menu"
                 else
                     mnu_open_menu "$_mp_label" "$_mp_key" "$menu"
@@ -523,7 +523,7 @@ menu_parse() {
                 ! tmux_vers_check "$_mp_min_vers" && continue
 
                 [ -n "$menu_debug" ] && debug_print "text line [$txt]"
-                if $cfg_use_whiptail; then
+                if $b_use_alt_handler; then
                     alt_text_line "$txt"
                 else
                     mnu_text_line "$txt"
@@ -539,7 +539,7 @@ menu_parse() {
                 [ -n "$menu_debug" ] && debug_print "Spacer line"
 
                 # Whiptail/dialog does not have a concept of spacer lines
-                if $cfg_use_whiptail; then
+                if $b_use_alt_handler; then
                     alt_spacer
                 else
                     mnu_spacer
@@ -597,7 +597,7 @@ menu_generate_part() {
 
     wt_actions=""
     menu_parse "$@"
-    $cfg_use_whiptail && update_wt_actions
+    $b_use_alt_handler && update_wt_actions
 }
 
 #---------------------------------------------------------------
@@ -673,7 +673,7 @@ check_screen_size() {
     #  including lines covered by a status bar. Since Menus can cover the status bar
     #  This gives the actual screen limits for menus
     #
-    $cfg_use_whiptail && return 0
+    $b_use_alt_handler && return 0
     # log_it "check_screen_size()"
 
     ${all_helpers_sourced:-false} || source_all_helpers "check_screen_size()"
@@ -743,7 +743,7 @@ set_menu_env_variables() {
 
     d_odd_chars="$d_items/odd_chars"
 
-    if [ "$cfg_use_whiptail" = true ]; then
+    if [ "$b_use_alt_handler" = true ]; then
         # Display Commands can only be used with tmux menus and caching
         cfg_display_cmds=false
         unset show_cmds_state
@@ -778,14 +778,14 @@ set_menu_env_variables() {
         #  items/main.sh -> cache/items/main.sh/
         d_menu_cache="$d_cache/$rn_current_script"
 
-        $cfg_use_whiptail && d_wt_actions="$d_menu_cache/wt_actions"
+        $b_use_alt_handler && d_wt_actions="$d_menu_cache/wt_actions"
     else
         uncached_menu=""
         uncached_wt_actions=""
         uncached_item_splitter="||||"
     fi
 
-    if $cfg_use_whiptail; then
+    if $b_use_alt_handler; then
         external_action_separator=":/:/:/:"
         #
         #  I haven't been able do to menu reload with whiptail/dialog yet,
@@ -1174,7 +1174,7 @@ display_menu() {
     # log_it "display_menu()"
     # Display time to generate menu
 
-    if $cfg_use_whiptail; then
+    if $b_use_alt_handler; then
         # display whiptail menu
         menu_selection=$(eval "$menu_items" 3>&2 2>&1 1>&3)
         menu_exit_code="$?"
@@ -1200,7 +1200,11 @@ display_menu() {
         eval "$menu_items" 2>"$f_cmd_err" || {
             display_invalid_menu_error "$_dm_err_msg"
         }
-        ensure_menu_fits_on_screen
+        tmux_vers_check 3.8 || [ "$current_tmux_vers" = "next-3.8" ] || {
+            # Pre-3.8: check if menu actually displayed
+            # TODO: Remove the next-3.8 check once tmux 3.8 is officially released
+            ensure_menu_fits_on_screen
+        }
     fi
 }
 
@@ -1220,8 +1224,6 @@ do_menu_handling() {
     #
     [ -z "$menu_name" ] && error_msg "menu_name not defined"
     [ -n "$menu_min_vers" ] && check_menu_min_vers
-    [ "$skip_oversized" = "1" ] && oversized_check
-
     menu_debug="" # Set to 1 to use echo 2 to use log_it
 
     prepare_menu